+++ apdt/plt/__init__.py
@@ -6,3 +6,4 @@ Provide useful data visualizaton tools.
 
 Check https://github.com/Zhiyuan-Wu/apdt for more information.
 '''
+from apdt.plt.logloader import logloader